requirements.txt
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.4
django-appconf==1.0.3
django-cleanup==4.0.0
django-cors-headers==3.1.0
django-environ==0.4.5
django-filter==2.2.0
django-googledrive-storage==1.3.5
django-rest-auth==0.9.5
djangorestframework==3.10.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==19.9.0
httplib2==0.14.0
httplib2shim==0.0.3
idna==2.8
oauth2client==4.1.3
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
psycopg2==2.8.3
